SMS sistem is now background service, have to find a way to get settings values to determine if it should even run. AND how to pass array prevozov to service????
