+++ app/src/main/AndroidManifest.xml
@@ -57,6 +57,10 @@
                 <action android:name="android.provider.Telephony.SMS_RECEIVED" />
             </intent-filter>
         </receiver>
+        <service
+            android:name=".SMSBackgroundService"
+            android:exported="false"/>
+
 
         <activity android:name=".LandingActivity"
             android:label="@string/title_activity_landing"
