+++ app/src/main/java/nameplaceholder/prevazanjaorg/SMSData.java
@@ -18,7 +18,8 @@ public class SMSData {
    public static final int SENT = 3;
    public static final int BAD_SMS = 4;
    public static final int PREKLIC = 5;
-   public static final int PRIVATE = 6;
+   public static final int STOP = -1;
+   public static final int START = 6;
 
 
     public SMSData(String s, String t){
