+++ app/src/main/res/layout/activity_list.xml
@@ -24,5 +24,6 @@
             android:divider="@android:color/transparent"
             android:dividerHeight="5dp"/>
 
+
     </RelativeLayout>
 </FrameLayout>
